[dev] fixed async connector + ssl, clean code, +macros, +cmake improvements

+++ include/qb/io/system/sys__socket.h
@@ -363,7 +363,7 @@ public:
 
     void
     af(int v) {
-        sa_.sa_family = v;
+        sa_.sa_family = static_cast<uint16_t>(v);
     }
     int
     af() const {
